conf/docker/twitter/requirements.txt
google-cloud-pubsub
tweepy
urllib3

